+++ packages/backend/src/services/UserService.ts
@@ -473,6 +473,9 @@ export class UserService extends BaseService {
         });
     }
 
+    /**
+     * @deprecated Group Sync is deprecated. https://github.com/lightdash/lightdash/issues/12430
+     */
     private async tryAddUserToGroups(
         ...data: ArgumentsOf<GroupsModel['addUserToGroupsIfExist']>
     ) {
@@ -1214,6 +1217,10 @@ export class UserService extends BaseService {
             }
             throw new AuthorizationError();
         }
+        // Update last used date
+        await this.personalAccessTokenModel.updateUsedDate(
+            personalAccessToken.uuid,
+        );
         return userWithOrganization;
     }