+++ packages/frontend/src/features/metricsCatalog/components/MetricsCatalogColumns.tsx
@@ -1,38 +1,97 @@
 import { type CatalogField } from '@lightdash/common';
-import { Box, Group, Highlight, HoverCard, Text, Tooltip } from '@mantine/core';
+import {
+    Flex,
+    Group,
+    Highlight,
+    HoverCard,
+    Text,
+    Tooltip,
+} from '@mantine/core';
+import { useHover } from '@mantine/hooks';
+import { IconPlus } from '@tabler/icons-react';
 import MarkdownPreview from '@uiw/react-markdown-preview';
 import { type MRT_ColumnDef } from 'mantine-react-table';
-import { useMemo } from 'react';
+import { useMemo, type FC, type SVGProps } from 'react';
+import MantineIcon from '../../../components/common/MantineIcon';
+import {
+    Description,
+    Hash,
+    Popularity,
+    Tag,
+} from '../../../svgs/metricsCatalog';
 import { useAppDispatch, useAppSelector } from '../../sqlRunner/store/hooks';
 import { setCategoryPopoverIsClosing } from '../store/metricsCatalogSlice';
 import { CatalogCategory } from './CatalogCategory';
+import { ExploreMetricButton } from './ExploreMetricButton';
 import { MetricChartUsageButton } from './MetricChartUsageButton';
 import { MetricsCatalogCategoryForm } from './MetricsCatalogCategoryForm';
 import { MetricsCatalogColumnName } from './MetricsCatalogColumnName';
 
+const HeaderCell = ({
+    children,
+    Icon,
+}: {
+    children: React.ReactNode;
+    Icon: FC<SVGProps<SVGSVGElement>>;
+}) => {
+    return (
+        <Group spacing={6} mr={6} h="100%" noWrap>
+            <Icon />
+            <Text
+                fz="xs"
+                fw={600}
+                color="dark.3"
+                sx={{
+                    // Turn off highlight text cursor - useful when resizing columns
+                    userSelect: 'none',
+                }}
+            >
+                {children}
+            </Text>
+        </Group>
+    );
+};
+
 export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
     {
         accessorKey: 'name',
-        header: 'Metric Name',
+        header: 'Metric',
         enableSorting: true,
         enableEditing: false,
-        Cell: ({ row, table }) => (
-            <Tooltip
-                label={row.original.tableName}
-                disabled={!row.original.tableName}
-                withinPortal
-                position="right"
-            >
-                <MetricsCatalogColumnName row={row} table={table} />
-            </Tooltip>
+        Header: ({ column }) => (
+            <HeaderCell Icon={Hash}>{column.columnDef.header}</HeaderCell>
         ),
+        Cell: ({ row, table }) => {
+            return (
+                <Flex justify="space-between" align="center" w="100%">
+                    <Tooltip
+                        label={row.original.tableName}
+                        disabled={!row.original.tableName}
+                        withinPortal
+                        position="right"
+                    >
+                        <MetricsCatalogColumnName row={row} table={table} />
+                    </Tooltip>
+                    <ExploreMetricButton
+                        row={row}
+                        className="explore-button"
+                        visibility="hidden"
+                    />
+                </Flex>
+            );
+        },
     },
     {
         accessorKey: 'description',
-        header: 'Description',
         enableSorting: false,
         enableEditing: false,
         size: 300,
+        header: 'Description',
+        Header: ({ column }) => (
+            <HeaderCell Icon={Description}>
+                {column.columnDef.header}
+            </HeaderCell>
+        ),
         Cell: ({ table, row }) => (
             <HoverCard
                 withinPortal
@@ -41,11 +100,18 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
                 disabled={!row.original.description}
             >
                 <HoverCard.Target>
-                    <Text lineClamp={2}>
+                    <Text
+                        lineClamp={2}
+                        c={row.original.description ? 'dark.4' : 'dark.1'}
+                        fz="sm"
+                        fw={400}
+                        lh="150%"
+                    >
                         <Highlight
                             highlight={table.getState().globalFilter || ''}
+                            lh="150%"
                         >
-                            {row.original.description ?? ''}
+                            {row.original.description ?? '-'}
                         </Highlight>
                     </Text>
                 </HoverCard.Target>
@@ -78,6 +144,9 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
                 },
             };
         },
+        Header: ({ column }) => (
+            <HeaderCell Icon={Tag}>{column.columnDef.header}</HeaderCell>
+        ),
         Edit: ({ table, row, cell }) => {
             const dispatch = useAppDispatch();
             const canManageTags = useAppSelector(
@@ -90,7 +159,7 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
             );
 
             return (
-                <Group spacing="two" pos="relative" w="100%" h="100%">
+                <Group spacing="xxs" pos="relative" w="100%" h="100%">
                     {categories.map((category) => (
                         <CatalogCategory
                             key={category.tagUuid}
@@ -121,6 +190,7 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
             );
         },
         Cell: ({ row, table, cell }) => {
+            const { hovered, ref } = useHover();
             const isCategoryPopoverClosing = useAppSelector(
                 (state) => state.metricsCatalog.popovers.category.isClosing,
             );
@@ -132,9 +202,11 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
 
             return (
                 // This is a hack to make the whole cell clickable and avoid race conditions with click outside events
-                <Box
+                <Flex
+                    ref={ref}
                     pos="absolute"
-                    p="md"
+                    py={6}
+                    px="md"
                     left={0}
                     top={0}
                     w="100%"
@@ -147,15 +219,28 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
                         table.setEditingCell(cell);
                     }}
                 >
-                    <Group spacing="two" pos="relative" w="100%" h="100%">
-                        {categories.map((category) => (
-                            <CatalogCategory
-                                key={category.tagUuid}
-                                category={category}
+                    {categories.length === 0 && hovered ? (
+                        <Group spacing={2}>
+                            <MantineIcon
+                                color="dark.1"
+                                icon={IconPlus}
+                                size={12}
                             />
-                        ))}
-                    </Group>
-                </Box>
+                            <Text span fz="sm" color="dark.1">
+                                Click to add
+                            </Text>
+                        </Group>
+                    ) : (
+                        <Group spacing="xxs" pos="relative" w="100%" h="100%">
+                            {categories.map((category) => (
+                                <CatalogCategory
+                                    key={category.tagUuid}
+                                    category={category}
+                                />
+                            ))}
+                        </Group>
+                    )}
+                </Flex>
             );
         },
     },
@@ -165,6 +250,16 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
         enableSorting: true,
         enableEditing: false,
         size: 100,
+        mantineTableBodyCellProps: () => {
+            return {
+                sx: {
+                    justifyContent: 'center',
+                },
+            };
+        },
+        Header: ({ column }) => (
+            <HeaderCell Icon={Popularity}>{column.columnDef.header}</HeaderCell>
+        ),
         Cell: ({ row }) => <MetricChartUsageButton row={row} />,
     },
 ];